requirements.txt
streamlit==1.41.1
requests==2.31.0
uuid==1.30
python-dotenv==1.0.0 